feat(mem,mt): rpc 传参改用确定性池化 sc_chunk/sc_recycle；run/sync/async 选项语法迁移
- op.h/op_impl.c：新增 sc_chunk/sc_recycle（恒走 mem 池，不受 SC_POOL 影响），
  与 SC_POOL 条件切换的 sc_alloc/sc_free 形成两层分配接口
- mt_impl.c：pool/queue/promise/thread 等全部 rpc 传参节点改用 sc_chunk/sc_recycle
- main.cpp：op 单元隐式硬依赖 mem，rpc 传参开箱即池化（无需 -DSC_POOL）
- run/sync/async <target, opt> 选项语法迁移（含 golden 快照更新）
- 回归 185/0，默认与 SC_POOL 构建运行时均验证通过

+++ tests/golden/feature41.c
@@ -59,9 +59,9 @@ int32_t main(void) {
     /* line 30 */
     queue *q = default_queue(p);
     /* line 33 */
-    promise *f1 = ({ struct compute _ap = {0}; _ap.a = 3; _ap.b = 4; q->async(q, (void (*)(void *))compute_rpc, &_ap, sizeof(_ap), 0, 0); });
+    promise *f1 = ({ struct compute _ap = {0}; _ap.a = 3; _ap.b = 4; q->async(q, (void (*)(void *))compute_rpc, &_ap, sizeof(_ap), (int32_t)0, (int64_t)0); });
     /* line 34 */
-    promise *f2 = ({ struct compute _ap = {0}; _ap.a = 100; _ap.b = 23; q->async(q, (void (*)(void *))compute_rpc, &_ap, sizeof(_ap), 0, 0); });
+    promise *f2 = ({ struct compute _ap = {0}; _ap.a = 100; _ap.b = 23; q->async(q, (void (*)(void *))compute_rpc, &_ap, sizeof(_ap), (int32_t)0, (int64_t)0); });
     /* line 35 */
     int32_t r1 = ((int32_t)(f1->wait(f1)));
     /* line 36 */
@@ -85,12 +85,12 @@ int32_t main(void) {
         struct consume_n _rp = {0};
         _rp.qq = q2;
         _rp.n = 2;
-        thread_run((void (*)(void *))consume_n_rpc, &_rp, sizeof(_rp), (thread **)(&(ct)), (uint32_t)0u, (uint8_t)0u);
+        thread_run((void (*)(void *))consume_n_rpc, &_rp, sizeof(_rp), (thread **)(&(ct)), (uint32_t)0, (uint8_t)0);
     }
     /* line 49 */
-    promise *g1 = ({ struct compute _ap = {0}; _ap.a = 10; _ap.b = 20; q2->async(q2, (void (*)(void *))compute_rpc, &_ap, sizeof(_ap), 0, 0); });
+    promise *g1 = ({ struct compute _ap = {0}; _ap.a = 10; _ap.b = 20; q2->async(q2, (void (*)(void *))compute_rpc, &_ap, sizeof(_ap), (int32_t)0, (int64_t)0); });
     /* line 50 */
-    promise *g2 = ({ struct compute _ap = {0}; _ap.a = 5; _ap.b = 6; q2->async(q2, (void (*)(void *))compute_rpc, &_ap, sizeof(_ap), 0, 0); });
+    promise *g2 = ({ struct compute _ap = {0}; _ap.a = 5; _ap.b = 6; q2->async(q2, (void (*)(void *))compute_rpc, &_ap, sizeof(_ap), (int32_t)0, (int64_t)0); });
     /* line 51 */
     int32_t s1 = ((int32_t)(g1->wait(g1)));
     /* line 52 */
